minio/requirements.txt
hydra-core==1.1.2
minio==7.1.6